+++ src/Resources/app/administration/src/module/rhae-tweakwise-frontend/index.js
@@ -11,7 +11,7 @@ Shopware.Module.register('rhae-tweakwise-frontend', {
     title: 'rhae-tweakwise-frontend.main.menuLabel',
     description: 'rhae-tweakwise-frontend.main.menuDescription',
     color: '#01b7cd',
-    icon: 'default-avatar-multiple',
+    icon: 'rhae-tweakwise-settings-icon',
 
     snippets: {
         'nl-NL': nlNL,