+++ src/Resources/app/administration/src/module/sw-cms/blocks/tweakwise/featured-products/preview/index.js
@@ -1,0 +1,6 @@
+import template from './sw-cms-preview-tweakwise-featured-products.html.twig';
+import './sw-cms-preview-tweakwise-featured-products.scss';
+
+Shopware.Component.register('sw-cms-preview-tweakwise-featured-products', {
+    template
+});