+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+### Added
+- You can now use the CMS element Tweakwise Featured Products in the Shopping Experience. With this element you can show featured products as specified in the Tweakwise App in your shop.
+
 ## [2.4.3] - 25-01-2024
 
 > ### Important